feat(go2_ws): Add the xt16 lidar link in the Go2 URDF model.
Set the argument `enable_xt16` to `True` if you want to use lidar.

Note that the transform from lidar base to lidar frame is computed from hesai_description, since we use `imu_link` as parent, we need to invert the transform in our URDF model. The transform from Go2 to lidar is copied from the Unitree docs.

Reference:
- https://github.com/ori-drs/hesai_description/blob/834ed126aef58bba51853c6fe900c808c5f6310a/urdf/hesai_xt32.urdf.xacro#L38-L40
- https://support.unitree.com/home/en/developer/SLAM%20and%20Navigation_service